support ssl flags reset
