+++ pkg/controller/component/helper.go
@@ -224,7 +224,7 @@ func staticOrStartupFlags(config map[string]string) map[string]string {
 	return static
 }
 
-func updateDynamicFlags(endpoints []string, newAnnotations map[string]string) error {
+func updateDynamicFlags(endpoints []string, oldAnnotations, newAnnotations map[string]string) error {
 	newFlags := make(map[string]string)
 	newFlagsVal, ok := newAnnotations[annotation.AnnLastAppliedDynamicFlagsKey]
 	if ok {
@@ -232,8 +232,20 @@ func updateDynamicFlags(endpoints []string, newAnnotations map[string]string) er
 			return err
 		}
 	}
-	if len(newFlags) == 0 {
+	oldFlags := make(map[string]string)
+	oldFlagsVal, ok := oldAnnotations[annotation.AnnLastAppliedDynamicFlagsKey]
+	if ok {
+		if err := json.Unmarshal([]byte(oldFlagsVal), &oldFlags); err != nil {
+			return err
+		}
+	}
+	if len(newFlags) == 0 && len(oldFlags) == 0 {
 		return nil
+	}
+	_, removed := maputil.IntersectionDifference(oldFlags, newFlags)
+	if len(removed) > 0 {
+		maputil.ResetMap(removed, v1alpha1.DynamicFlags)
+		newFlags = maputil.MergeStringMaps(true, newFlags, removed)
 	}
 	klog.V(1).Infof("dynamic flags: %v", newFlags)
 	str, err := codec.Encode(newFlags)
